Changed spheres to cubes

+++ Assets/Shaders/Unlit/TexturedShader.shader
@@ -44,7 +44,7 @@ Shader "Unlit/TexturedShader"
 
             fixed4 frag (v2f i) : SV_Target
             {
-                fixed4 col = tex2D(_MainTex, i.uv);
+                float4 col = tex2D(_MainTex, i.uv);
                 return col;
             }
             ENDCG
